move some luagamestate_support functions into luaframework
