+++ tools/code_generators/luagamestate_support_gen/main.cpp
@@ -92,6 +92,7 @@ void writeSource(std::ofstream &out, const StateDefinition &state,
                  const std::string &outputHeaderFile)
 {
 	out << "// GENERATED SOURCE - generated by LuaGamestateSupportGen - do not modify directly\n\n";
+	out << "#include \"framework/luaframework.h\"\n";
 	out << "#include \"game/state/luagamestate_support.h\"\n";
 	out << "#include \"" << outputHeaderFile << "\"\n\n";
 
@@ -224,7 +225,7 @@ void writeSource(std::ofstream &out, const StateDefinition &state,
 		    << " v) { lua_pushinteger(L, static_cast<int>(v)); }\n";
 	}
 
-	out << "void pushLuaEnums(lua_State *L)\n"
+	out << "void pushLuaGamestateEnums(lua_State *L)\n"
 	    << "{\n"
 	    << "\tlua_createtable(L, 0, 0);\n";
